ENGR00132800 SRTC Time set notification test fails with SD fs
The child process was preventing the parent from running
initially to enter wait queue for notification. Therefore,
subsequent time change by the child process was not being caught
by the parent. Implemented the following changes to the test case:

1) Decreased priority of the child (15) in comparison with parent (0).
This ensures that parent can process the time change right after child
sets the new time.

2) The above priority change was not enough to yield to the parent at
the beginning of the test for it to complete its ioctl call to enter
wait queue. So child process needs to sleep for 1 second to allow
parent to complete the wait queue enter call before time is changed.

Signed-off-by: Anish Trivedi <anish@freescale.com>