ENGR00331509 mxc_alstest: Get the input event index correctly
Before the 3.14 kernel , we can use the input index in "/sys/devices/virtual/input/inputx"
as event index in "/dev/input/eventx", because they are the same.
But in the 3.14 kernel, the index doesn't keep the same all the time. So we need to search the
folder of "/sys/devices/virtual/input/inputx" to get the related event index.

Signed-off-by: Luwei Zhou <b45643@freescale.com>