MLK-14294 mxc_uart_stress_test: validate chunk_size argument
If CHUNK_SIZE argument is higher than 1024, we will run into a
buffer overflow issue. The buffer used for sending and receiving
data is statically allocated at 1024 characters (the default
CHUNK_SIZE).

Remove the hardcoded value of 1024 and replace it with a constant.
Check the CHUNK_SIZE specified as program argument and make sure
it is lower than the default.

Signed-off-by: Irina Tirdea <irina.tirdea@nxp.com>