MLK-14661: pxp: v4l2_capture: check disk writes
During the capture test, when saving the frames to disk (using fwrite),
no check was done. Therefore, even if the current file system is full,
the frames were "written" without checking for errors, thus the actual
frames were lost.
Add a check to the fwrite function and stop when the written data is
less than expected.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>