MLK-14719: mxc_fb_test: Add test: fb_test_images
Added a new test fb_test_images. This test will use image files on
current directory of the format <name>-<xres>x<yres>-<bpp>.rgb in order
to fill the framebuffers. Depending on the image size, it will use BG or
FG framebuffer. This test is useful to test different resolutions and
bit-per-pixel on the framebuffers.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>