MLK-15049-3 test: add basic acceptance tests
These tests are designed to be run automatically and with a fast
response time. The rule of thumb is that a single test should not take
more than 5 seconds.

The tests should only check for basic functionality and should not
contain any stress tests or long duration tests.

The test output format is TAP[1] and since this is used by tools it
should be used consistently for every new test added. The bat.sh
script takes care of properly formatting the output but each
individual test must return the proper value: 0 for success, 1 for
failure and BAT_EXITCODE_SKIP if the test was skipped on a particular
platform.

[1] https://testanything.org/

Signed-off-by: Octavian Purdila <octavian.purdila@nxp.com>
Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Signed-off-by: Irina Tirdea <irina.tirdea@nxp.com>
Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>