MLK-14764-2: bat: Add playback/capture tests to audio BAT
For every audio cards found in a board run a simple playback or capture
test to check its basic functionality. Because these tests are executed
by CI for every new patch we want them finish as fast as possible so
we limit a test for a device to run only two scenarios:
1. one second, first supported rate, format and channels
2. one second, last supported rate, format and channels

For now HDMI audio cards are skipped because we do not have HDMI
devices attached to the boards under test.

Signed-off-by: Ion-Horia Petrisor <ion-horia.petrisor@nxp.com>
Signed-off-by: Mihai Serban <mihai.serban@nxp.com>
Reviewed-by: Daniel Baluta <daniel.baluta@nxp.com>