MLK-16116: bat-cpufreq: Set cpufreq governor to ondemand
This test expects cpu to be set to a low frequency when not
performing an intensive task and on high frequency whet it does.

This behaviour requires the cpufreq governor to be set to "ondemand".
Patch sets governor to ondemand during the test and restores it after
(similar to cpufreq-freq-check test).

Signed-off-by: Adriana Reus <adriana.reus@nxp.com>
Reviewed-by: Leonard Crestez <leonard.crestez@nxp.com>