MLK-16185-1: bat: Add support for TODO tests
Scripts that test a not yet implemented feature or not yet fixed bug
should exit with BAT_EXITCODE_TODO instead of BAT_EXITCODE_SKIP.

According to TAP13 specification the tests marked with TODO directive
represents a feature to be implemented or a bug to be fixed and act as
something of an product 'things to to' list. They are not expected to
succeed. Should a todo test point begin succeeding, the harness should
report it as a bonus. This indicates that whatever you were supposed
to do has been done and you should promote this to a normal test point.

https://testanything.org/tap-version-13-specification.html

Signed-off-by: Mihai Serban <mihai.serban@nxp.com>
Reviewed-by: Leonard Crestez <leonard.crestez@nxp.com>