MLK-14983: bat_utils.sh: Use read -r to read current freq
Upstream wants to clamp the ondemand sampling frequency to 10ms max but
that interval is so short and our CPUs are so slow that merely executing
"cat sampling_cur_freq" will result in increasing the frequency before
it returns. This causes the test to fail because the frequency read back
is always the maximum.

Fix this by using "read" instead of cat, running a shell builtin is many
times faster that executing a new process.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Reviewed-by: Mihai Serban <mihai.serban@nxp.com>