bat-uart.sh: Fix pausing/stopping multiple processes
Attempting to run bat-uart.sh through ssh while logged in on the
serial console results in "invalid signal specification". This happens
because when multiple processes use the same console the
signal_processes function is called with multiple pids and improper
quoting: the second pid is interpreted as "signal number"

The kill command already supports a list of pids so just use that
directly.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Reviewed-by: Raluca Oncioiu <raluca.oncioiu@nxp.com>
Reviewed-by: Daniel Baluta <daniel.baluta@nxp.com>