MLK-18567 bat: Unbind imx_usb driver when entering busfreq
This is done from the same code which handles network down/restore
because network might be on USB and this needs special handling.

There are no cases where we need to unbind/rebind usb other than
busfreq.

For usb network the device gets unregistered so we need to wait for it
to probe and assign the IP address again.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Reviewed-by: Daniel Baluta <daniel.baluta@nxp.com>