MLK-19247: mxc-jpeg: Remove jpeg-encoder hardcoding to 64x64 yuv422
Encoder test application needs to feed the wxh and
pixel format to the driver, since it cannot be extracted
from the raw file.
Decoder test application has some information in the
jpeg markers, but YUV444 and RGB24 have the same signature
in the SOH marker, because they have the same subsampling
scheme.
So, both the decoder and encoder  test application will take
<width>x<height> and pixel format from the user.

Fixed the size of the resulting jpeg buffer, as it had
some trailing garbage at the end.

Signed-off-by: Mirela Rabulea <mirela.rabulea@nxp.com>