bat_utils: Run everything if SOC version can't be determined
We currently don't have a SOC driver in upstream for any imx8. BAT will
try to skip unsupported tests on a per-SOC basis but if the SOC can't be
determined it fails and doesn't run anything.

Fix this by running all tests if no soc version info can be found in
sysfs. With this patch bat-cpufreq passes on imx8mq upstream

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Reviewed-by: Daniel Baluta <daniel.baluta@nxp.com>