MLK-23069: mxc_alsa_conf: Add script to update asound.conf
In asound.conf, we set the "hw:0,0" as the default sound
card device. But "hw:0,0" may not be a proper device for
listening sound or recording sound data.

In Linux Factory kernel there is "defer probe" for modules,
that cause the "hw:0,0" maybe "imx-audmix" sound card, which
is not a traditional sound card device.

So we write this script, which will search wm8960/wm8962/wm8524/
cs42888, update their sound index to /etc/asound.conf, user
then can use "aplay xxx.wav" to hear the sound.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>