MLK-25763: mxc_v4l2_vpu_test: open libg2d using dllopen instead of linking it
there is no libg2d.so on imx8mq,
so if the mxc_v4l2_vpu_test link it,
the mxc_v4l2_vpu_test can't be executed on imx8mq.
As the libg2d is different on imx8q and imx8m,
so we can't use static link.
mxc_v4l2_vpu_test will load the libg2d.so dynamically

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Shijie Qin <shijie.qin@nxp.com>