MLK-25799-3: mxc_v4l2_vpu_test: ignore pollerr on decoder's output queue
v4l2_m2m_fop_poll will return pollerr if capture is not streamon,
or capture's queue is empty,
but for decoder, unittest will wait source change event before setup
capture, so it will always get a pollerr event except driver drop it.
so in unittest we can ignore it on output queue,
if there is indeed some error occurs, we get receive pollerr on capture
queue, then we can handle it.

Signed-off-by: Ming Qian <ming.qian@nxp.com>