LF-5669: mxc-jpeg: Switch from RGB/ARGB to BGR/ABGR
The hardware is capable of encoding/decoding RGB and ARGB formats in
whatever order the color components are, but the resulting jpegs look
good if we start with raw data in BGR/ABGR order, so the driver was
changed to support only V4L2_PIX_FMT_BGR24 and V4L2_PIX_FMT_ABGR32.

Keep the RGB/ARBG formats as obsolete, for compatibility with
older versions of the driver.

Signed-off-by: Mirela Rabulea <mirela.rabulea@oss.nxp.com>
Reviewed-by: ming_qian <ming.qian@nxp.com>