LF-5444: mxc-jpeg: Add parameter for quality factor
This applies to the encoder only, the quality factor
may be in 1..100 range, where larger values correspond to
better image quality.
With a quality factor of 50, the DQT from the driver
will be used unmodified, the same should happen with factor 0,
but, per V4L2 spec, a quality factor of 0 would imply quantization
tables are configured directly by user application, which the
driver does not support, so we will not support factor 0.

Signed-off-by: Mirela Rabulea <mirela.rabulea@oss.nxp.com>