LF-6065-2: mxc_v4l2_vpu_test: mxc_v4l2_vpu_dec.out handle pollerr
When vpu driver report error,
driver will return pollerr in poll function,
and trigger an error event,
but the error event is custom, maybe drop in future.

In mxc_v4l2_vpu_dec.out, it doesn't handle pollerr,
and it doesn't open video device with O_NONBLOCK flag.
In select(), the pollerr will be treated as readable, and writable.
So the unitest may be blocked at dqbuf.

Signed-off-by: Ming Qian <ming.qian@nxp.com>