LF-11291: mxc_v4l2_vpu_test: connect decoder and its sink in source change
mxc_v4l2_vpu_test may connect the decoder and its seek after source
change event is received, but if the first frame is decoded in the same
time, then the first frame may be dropped as the sink is not connected.

connect the decoder and sink in handling source change event, that can
avoid such frame dropped issue.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: TaoJiang <tao.jiang_2@nxp.com>